Readability improvement and advances on presented groups.
Changed implementation of permutations from array to association list.
This means that types to be permuted over now have to be instances of
Enum instead of Ord and Ix, which is slightly cleaner and doesn't
require users to import Data.Array. Also, it should make the code easier
to understand for people with limited Haskell knowledge, as only basic
types are used.

Renamed PresentedGroupElement to Word, which is shorter and more
descriptive.

applyRewriteRules rewritten to actually work correctly and to be
(hopefully) easier to understand (although it contains a do block using
the Maybe monad - but presumably one does not have to grok monads in
order to understand the code). In any case, the whole implementation of
presented groups will be considerably more involved than that of
permutation groups; anyone interested only in the basics can stick with
the latter, while anyone wanting to understand the former will have to
invest a bit more effort anyway owing to the inherent complexity of the
subject.