Improved handling of presentions.
Handling of free and finitely presented groups should now be closer to
the mathematical concepts and considerably easier to use. Relators can
now be constructed without knowledge of the internal representation by
virtue of a separete Group instance for the free group generated by
values of some type and by using the gen function for embedding values
of this type into the free group (as words of length one).